requirements.txt
Flask
python-telegram-bot==22.5
requests
beautifulsoup4
gunicorn